fix(ui): move dark mode toggle to home page next to help button
The theme toggle in the Layout header was taking vertical space and
pushing all content down. Moved it into the Home page toolbar alongside
the help button, removing the now-empty header from Layout.
